feat(selected_line): plot selected lines in plots
For the axial view, plot the selected axial value as a horizontal dashed
line.

For the time plot, plot the selected time as a vertical dashed line.

This commit also fixes an issue where we were relying on state-listening functions
to be called in a certain order for the selected time to update properly. We no longer do that.

Signed-off-by: Patrick Avery <patrick.avery@kitware.com>